new file:   .vscode/settings.json 	modified:   client/src/app.css 	new file:   client/src/components/Editor.jsx 	modified:   client/src/components/Post.jsx 	modified:   client/src/components/index.js 	modified:   client/src/pages/Article.jsx 	modified:   client/src/pages/CreatePost.jsx 	modified:   client/src/pages/Posts.jsx 	modified:   client/src/userContext.js 	modified:   server/.gitignore 	modified:   server/controllers/postControllers.js 	modified:   server/index.js 	modified:   server/models/post.js 	modified:   server/package-lock.json 	modified:   server/package.json 	modified:   server/routes/postRoutes.js

+++ server/controllers/postControllers.js
@@ -1,8 +1,10 @@
 import mongoose from "mongoose";
 import PostMessage from "../models/post.js";
-import cloudinary from "../utils/cloudinary.js";
+// import cloudinary from "../utils/cloudinary.js";
 import Jwt from 'jsonwebtoken';
 import dotenv from 'dotenv';
+import fs from 'fs';
+
 
 dotenv.config();
 
@@ -27,7 +29,6 @@ export const getPosts = async(req, res) => {
     try {
 
         const data = await PostMessage.find().populate('author', ['username']).sort({ createdAt: -1 });
-        // console.log(data)
 
         res.status(200).json(data);
         
@@ -38,27 +39,27 @@ export const getPosts = async(req, res) => {
 }
 
 export const createPost = async(req, res) => {
-    const { token } = req.cookies;
-    // console.log(token)
-    const { title, summary, img, content } = req.body;
+    
     try {
+        const { originalname, path } = req.file;
+        const ext = originalname.split('.')[originalname.split('.').length -1];
+        const newName = `${path}.${ext}`;
+        fs.renameSync(path, newName);
+
+        const { title, summary, content } = req.body;
+        const { token } = req.cookies;
 
         Jwt.verify(token, process.env.JWT_SECRET, {}, async(err, info)=> {
-            // console.log(info._id);
             if (err) return res.status(400).json({ message: 'Invalid token' });
-            if(img) {
+            
+            if(newName){
     
-                const uploadRes = await cloudinary.uploader.upload(img, { upload_preset: 'erics_blog_photos'});
-    console.log(uploadRes)
-                if(uploadRes){
+                const newPost = await new PostMessage({ title, summary, img: newName, content, author: info._id });
         
-                    const newPost = await new PostMessage({ title, summary, img: uploadRes, content, author: info._id });
-            
-                    await newPost.save();
-            
-                    res.status(200).json(newPost);
-                }
-            }
+                await newPost.save();
+        
+                res.status(200).json(newPost);
+            }            
             
         })
 
@@ -69,27 +70,40 @@ export const createPost = async(req, res) => {
 }
 
 export const updatePost = async(req, res) => {
-    const post = req.body;
-    const { img } = post;
+    let newName;
+
+    if (req.file) {
+        
+        const { originalname, path } = req.file;
+        const ext = originalname.split('.')[originalname.split('.').length - 1];
+        newName = `${path}.${ext}`;
+        fs.renameSync(path, newName);
+    
+    }
+
+    const {title, summary, content } = req.body;
+    const { token } = req.cookies;
     const { id: _id } = req.params;
 
     if(!mongoose.Types.ObjectId.isValid(_id)) return res.status(400).json({ message: 'Invalid id' });
-            
+    const postData = await PostMessage.findById( _id );
+    
     try {
-
-        if(img) {
+        Jwt.verify(token, process.env.JWT_SECRET, {}, async(err, info)=> {
+            if (err) return res.status(400).json({ message: 'Invalid token' });
+            const isAuthor = JSON.stringify(postData.author) === JSON.stringify(info._id)
             
-            const uploadRes = await cloudinary.uploader.upload(img, { upload_preset: 'erics_blog_photos'});
-            console.log('image uploaded')
+            
 
-            if(uploadRes) {
+             if(isAuthor) {
 
-                const updatedPost = await PostMessage.findByIdAndUpdate( _id, { ...post, img: uploadRes, _id }, { new: true });
-                
-                console.log('success')
-                res.status(200).json(updatedPost);
-            }
-        }
+                    const updatedPost = await PostMessage.findByIdAndUpdate( _id, { title, summary, content, img: newName ? newName : postData.img, _id }, { new: true });
+
+                    res.status(200).json(updatedPost);
+                } else {
+                    res.status(403).json({ message: 'You are not allowed to edit this post. Only the author can edit' });
+                }
+        });
         
     } catch (error) {
         console.error( error.message || error );
@@ -100,12 +114,28 @@ export const updatePost = async(req, res) => {
 
 export const deletePost = async(req, res) => {
     const { id: _id } = req.params;
+    const { token } = req.cookies;
     
     try {
-    
         if(!mongoose.Types.ObjectId.isValid(_id)) return res.status(400).json({ message: 'Invalid id' });
+        
+        Jwt.verify( token, process.env.JWT_SECRET, {}, async(err, info) => {
 
-        const success = await PostMessage.findByIdAndDelete( _id );
+            if(err) return res.status(403).json({ message: `You are not authorized to delete this Article` });
+
+            const postData = await PostMessage.findById( _id );
+
+            const isAuthor = JSON.stringify(postData.author) === JSON.stringify(info._id);
+
+            if (!isAuthor) return res.status(403).json({ message: `You are not authorized to delete this Article, Only the author can delete` });
+
+            if (isAuthor) {
+
+                const success = await PostMessage.findByIdAndDelete( _id );
+            }
+
+        })
+
 
 
         res.status(200).json({ message: success });
